backend_project/requirements.txt
Django==3.2
djangorestframework>=3.11.0,<3.12.0
django-cors-headers==4.3.1
dj-database-url==0.5.0
psycopg2~=2.9.2
